requirements.txt
transformers==4.51.3
pillow==11.2.1
scipy==1.15.0
tqdm==4.67.1
pytorchvideo
eva-decord
opencv-python
numpy<2.0.0
dotenv
torch
torchvision==0.17.2
scikit-learn